Use Dotnet.Testcontainers
This commit uses DotNet.Testcontainers instead
of TestEnvironment.Docker. Version 1.0.0 is used
as support for net461 is needed.
