--- test/Elastic.Apm.DatabaseTests.Common/Elastic.Apm.DatabaseTests.Common.projitems
@@ -1,14 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>94100523-ee50-4912-873e-d1a15813a2b6</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>Elastic.Apm.DatabaseTests.Common</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)DatabaseFixture.cs" />
-  </ItemGroup>
-</Project>
