+++ test/Elastic.Apm.SqlClient.Tests/EfCoreWithMsSqlTests.cs
@@ -5,7 +5,6 @@
 using System;
 using System.Diagnostics;
 using System.Linq;
-using Elastic.Apm.DatabaseTests.Common;
 using Elastic.Apm.EntityFrameworkCore;
 using Elastic.Apm.Tests.Utilities;
 using FluentAssertions;
