+++ test/Elastic.Apm.SqlClient.Tests/Elastic.Apm.SqlClient.Tests.csproj
@@ -18,8 +18,9 @@
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.8.3" />
     <PackageReference Include="StrongNamer" Version="0.2.5" />
     <PackageReference Include="System.Data.SqlClient" Version="4.7.0" />
-    <PackageReference Include="TestEnvironment.Docker" Version="1.3.9" />
-    <PackageReference Include="TestEnvironment.Docker.Containers.Mssql" Version="1.1.0" />
+    <PackageReference Include="DotNet.Testcontainers" Version="1.0.0" />
+    <!-- Strong name any assemblies that aren't i.e. DotNet.Testcontainers -->
+    <PackageReference Include="StrongNamer" Version="0.2.5" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
       <PrivateAssets>all</PrivateAssets>
@@ -34,7 +35,5 @@
     <ProjectReference Include="..\..\src\Elastic.Apm.SqlClient\Elastic.Apm.SqlClient.csproj" />
     <ProjectReference Include="..\Elastic.Apm.Tests.Utilities\Elastic.Apm.Tests.Utilities.csproj" />
   </ItemGroup>
-
-  <Import Project="..\Elastic.Apm.DatabaseTests.Common\Elastic.Apm.DatabaseTests.Common.projitems" Label="Shared" />
 
 </Project>
