+++ test/Elastic.Apm.SqlClient.Tests/SqlServerCollection.cs
@@ -1,0 +1,14 @@
+// Licensed to Elasticsearch B.V under
+// one or more agreements.
+// Elasticsearch B.V licenses this file to you under the Apache 2.0 License.
+// See the LICENSE file in the project root for more information
+
+using Xunit;
+
+namespace Elastic.Apm.SqlClient.Tests
+{
+	[CollectionDefinition("SqlServer")]
+	public class SqlServerCollection : ICollectionFixture<SqlServerFixture>
+	{
+	}
+}
